+++ onboarding-api/src/utils/types.d.ts
@@ -1,3 +1,5 @@
+import { SupportedNetworks } from '../database'
+
 export type Subset<K> = {
   [attr in keyof K]?: K[attr] extends object
     ? Subset<K[attr]>
@@ -33,35 +35,6 @@ declare global {
       PURE_PROXY_CONTROLLER_SEED: string
       ONBOARDING_STORAGE_BUCKET: string
       EVM_ON_SUBSTRATE_CHAIN_ID: string
-    }
-  }
-}
-
-type VerificationState = 1 | 0 | null
-
-export type ManualKybCallbackRequestBody = {
-  reference: `MANUAL_KYB_REFERENCE_${string}`
-  event:
-    | `request.${'pending' | 'timeout' | 'deleted' | 'received'}`
-    | 'review.pending'
-    | `verification.${'accepted' | 'declined' | 'cancelled' | 'status.changed'}`
-  verification_url: `https://app.shuftipro.com/verification/process/${string}`
-  email: string
-  country: string
-
-  /**
-   * This object will be returned in case of verification.accepted or verification.declined.
-   * This object will include all the gathered data in a request process.
-   */
-  verification_data?: unknown
-  verification_result?: {
-    proof_stores: {
-      articles_of_association: VerificationState
-      certificate_of_incorporation: VerificationState
-      proof_of_address: VerificationState
-      register_of_directors: VerificationState
-      register_of_shareholders: VerificationState
-      signed_and_dated_ownership_structure: VerificationState
     }
   }
 }
